requirements.txt
dspy-ai>=2.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
requests>=2.31.0
pytest>=7.4.0
openai>=1.0.0